refactor out map
