+++ src/App.tsx
@@ -1,8 +1,5 @@
 import { Card, ConfigProvider, Splitter } from 'antd';
-import { PathOptions, StyleFunction, CircleMarker, LatLngExpression } from 'leaflet'
 import './App.css'
-import { MapContainer, Marker, Popup, GeoJSON, TileLayer } from 'react-leaflet'
-import { Feature, Geometry} from 'geojson'
 import { useEffect, useRef, useState } from 'react'
 import Layers from './components/Layers.tsx';
 import Properties from './components/Properties.tsx';
@@ -14,30 +11,11 @@ import track from './data/track1.ts';
 import track2 from './data/track2.ts';
 import zones from './data/zones.ts';
 import points from './data/points.ts';
-
-interface CustomPathOptions extends PathOptions {
-  radius?: number;
-}
-
-const setColor: StyleFunction = (feature: Feature<Geometry, unknown> | undefined) => {
-  const res: CustomPathOptions = {}
-  if (feature) {
-    const feat = feature as Feature
-    if (feat?.properties?.color) {
-      res.color = feat.properties.color
-    }
-    if (feat.geometry.type === 'Point') {
-      res.radius = 30
-    }
-  }
-  res.weight = 3
-  return res;
-};
+import Map from './components/Map.tsx';
 
 function App() {
   const features = useAppSelector(state => state.featureCollection.features)
   const dispatch = useAppDispatch()
-  const [selectedFeature, setSelectedFeature] = useState<Feature | null>(null);
   const [timeBounds, setTimeBounds] = useState<[number, number]>([0, 0])
 
   const initialised = useRef(false); 
@@ -57,11 +35,6 @@ function App() {
     }
   }, [dispatch])
 
-  const createLabelledPoint = (pointFeature: Feature, latlng: LatLngExpression) => {
-    const color = pointFeature.properties?.color || 'blue';
-    const name = pointFeature.properties?.name || '';
-    return new CircleMarker(latlng, { radius: 15, color }).bindTooltip(name, { permanent: true, direction: 'center' });
-  }
 
   const antdTheme = {
     components: {
@@ -75,18 +48,10 @@ function App() {
     }
   }
 
-  const setSelected = (ids: string[]): void => {
-    const selected = features.find((feature) => ids.includes(feature.id as string)) || null;
-    setSelectedFeature(selected);
-  }
-
   const setTime = (value: number) => {
     console.log('new time:', value, new Date(value).toISOString())
   }
 
-  const isVisible = (feature: Feature): boolean => {
-    return feature.properties?.visible
-  }
 
   return (
     <div className="App">
@@ -114,19 +79,7 @@ function App() {
             </Splitter>
           </Splitter.Panel>
           <Splitter.Panel key='right'>
-            <MapContainer center={[35.505, -4.09]} zoom={8} scrollWheelZoom={true}>
-              <TileLayer attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors' 
-                url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"/>
-              { 
-                features.filter(feature => isVisible(feature)).map((feature, index) => 
-                  <GeoJSON key={`${feature.id || index}`} data={feature} style={setColor} pointToLayer={createLabelledPoint}/>)
-              }
-              <Marker position={[51.505, -0.09]}>
-                <Popup>
-                A pretty CSS3 popup. <br /> Easily customizable.
-                </Popup>
-              </Marker>
-            </MapContainer>
+            <Map />
           </Splitter.Panel>
         </Splitter>
       </ConfigProvider>
